refactor: extract endorsement request logic into utils module
Move endorsement request functionality from resource.py to endorsement_request_utils.py for better separation of concerns.
